requirements.txt
pyserial==3.5
PySide6==6.6.2
PySide6_Addons==6.6.2
PySide6_Essentials==6.6.2
python-dotenv==1.0.1
shiboken6==6.6.2
